+++ dds/src/dds/topic_definition/type_support.rs
@@ -14,7 +14,7 @@ use crate::{
         instance::InstanceHandle,
     },
     serialized_payload::{
-        deserialize::CdrDeserialize, parameter_list_deserialize::ParameterListDeserialize,
+        cdr::deserialize::CdrDeserialize, parameter_list_deserialize::ParameterListDeserialize,
         parameter_list_serialize::ParameterListSerialize,
         parameter_list_serializer::ParameterListSerializer, serialize::CdrSerialize,
     },
