+++ dds/src/implementation/data_representation_inline_qos/types.rs
@@ -1,4 +1,4 @@
-use crate::serialized_payload::{deserialize::CdrDeserialize, serialize::CdrSerialize};
+use crate::serialized_payload::{cdr::deserialize::CdrDeserialize, serialize::CdrSerialize};
 
 #[derive(Clone, Copy, CdrSerialize, CdrDeserialize, PartialEq, Eq)]
 pub struct KeyHash(pub [u8; 16]);
