+++ dds/src/implementation/payload_serializer_deserializer/cdr_deserializer.rs
@@ -4,7 +4,7 @@ use std::{
 };
 
 use crate::serialized_payload::{
-    deserialize::CdrDeserialize, deserializer::CdrDeserializer, error::CdrResult,
+    cdr::deserialize::CdrDeserialize, cdr::deserializer::CdrDeserializer, error::CdrResult,
 };
 
 use super::endianness::CdrEndianness;
