+++ dds/src/serialized_payload/mod.rs
@@ -1,6 +1,6 @@
-pub mod deserialize;
-pub mod deserializer;
+pub mod cdr;
 pub mod error;
+pub mod parameter_list;
 pub mod parameter_list_deserialize;
 pub mod parameter_list_deserializer;
 pub mod parameter_list_serialize;
